backend/requirements.txt
fastapi
uvicorn
python-multipart
onnxruntime
numpy
Pillow
